♻️ (donate) externalize constants
