+++ functions/donation/donate.ts
@@ -8,6 +8,7 @@ import {
     stringifyUnknownError,
 } from "@ourworldindata/utils"
 import { Value } from "@sinclair/typebox/value"
+import { DEFAULT_HEADERS, CORS_HEADERS } from "./_utils/constants.js"
 
 interface DonateEnvVars {
     ASSETS: Fetcher
@@ -17,24 +18,6 @@ interface DonateEnvVars {
 
 const hasDonateEnvVars = (env: any): env is DonateEnvVars => {
     return !!env.ASSETS && !!env.STRIPE_SECRET_KEY && !!env.RECAPTCHA_SECRET_KEY
-}
-
-// CORS headers need to be sent in responses to both preflight ("OPTIONS") and
-// actual requests.
-const CORS_HEADERS = {
-    "Access-Control-Allow-Origin": "*",
-    "Access-Control-Allow-Methods": "POST, OPTIONS",
-    // The Content-Type header is required to allow requests to be sent with a
-    // Content-Type of "application/json". This is because "application/json" is
-    // not an allowed value for Content-Type to be considered a CORS-safelisted
-    // header.
-    // - https://developer.mozilla.org/en-US/docs/Glossary/CORS-safelisted_request_header
-    "Access-Control-Allow-Headers": "Content-Type",
-}
-
-export const DEFAULT_HEADERS = {
-    ...CORS_HEADERS,
-    "Content-Type": "application/json",
 }
 
 // This function is called when the request is a preflight request ("OPTIONS").
